backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
flask-bcrypt==1.0.1
pyjwt==2.8.0
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
PyMySQL==1.1.1
cryptography==42.0.8
